+++ ReleaseNotes.md
@@ -14,6 +14,7 @@ Migrated bundles from under `packages` to `bundles` and/or to new bundle-loader 
 
 - packages/admin/bundle/admin-announcements/bundle.js -> bundles/admin/admin-announcements/
 - packages/admin/bundle/admin-layereditor/bundle.js -> bundles/admin/admin-layereditor/
+- packages/admin/bundle/admin-permissions/bundle.js -> bundles/admin/admin-permissions/
 
 
 ### Changes for bundle registrations